server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pdfminer.six==20221105
werkzeug==3.0.1
python-dotenv==1.0.0